show class members
