+++ noval/parser/config.py
@@ -17,7 +17,7 @@ ASSIGN_TYPE_INT = 0
 ASSIGN_TYPE_STR = 1
 ASSIGN_TYPE_MODULE = 2
 ASSIGN_TYPE_LIST =3
-ASSIGN_TYPE_TUPLE = 3
+ASSIGN_TYPE_TUPLE = 4
 ASSIGN_TYPE_DICT = 5
 ASSIGN_TYPE_OBJECT = 6
 ASSIGN_TYPE_FLOAT = 7
